requirements.txt
fastapi==0.103.2
uvicorn==0.24.0
watchdog==3.0.0
pytesseract==0.3.10
pdf2image==1.16.3
pdfplumber==0.10.2
pymupdf==1.23.7
python-multipart==0.0.6
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==1.10.13
httpx==0.25.1
python-dotenv==1.0.0
pyyaml==6.0.1
greenlet==3.0.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
ics==0.7.2
bcrypt==3.2.2
